+++ lua/lazyvim/plugins/editor.lua
@@ -176,12 +176,15 @@ return {
     event = "VeryLazy",
     opts_extend = { "spec" },
     opts = {
+      preset = "helix",
       defaults = {},
       spec = {
         {
           mode = { "n", "v" },
           { "<leader><tab>", group = "tabs" },
           { "<leader>c", group = "code" },
+          { "<leader>d", group = "debug" },
+          { "<leader>dp", group = "profiler" },
           { "<leader>f", group = "file/find" },
           { "<leader>g", group = "git" },
           { "<leader>gh", group = "hunks" },
@@ -300,6 +303,20 @@ return {
       end,
     },
   },
+  {
+    "gitsigns.nvim",
+    opts = function()
+      Snacks.toggle({
+        name = "Git Signs",
+        get = function()
+          return require("gitsigns.config").config.signcolumn
+        end,
+        set = function(state)
+          require("gitsigns").toggle_signs(state)
+        end,
+      }):map("<leader>uG")
+    end,
+  },
 
   -- better diagnostics list and others
   {
@@ -378,6 +395,12 @@ return {
     import = "lazyvim.plugins.extras.editor.telescope",
     enabled = function()
       return LazyVim.pick.want() == "telescope"
+    end,
+  },
+  {
+    import = "lazyvim.plugins.extras.editor.snacks_picker",
+    enabled = function()
+      return LazyVim.pick.want() == "snacks"
     end,
   },
 }