+++ lua/lazyvim/plugins/extras/lang/sql.lua
@@ -48,7 +48,7 @@ return {
       vim.api.nvim_create_autocmd("FileType", {
         pattern = sql_ft,
         callback = function()
-          if LazyVim.has("nvim-cmp") then
+          if LazyVim.cmp_engine() == "nvim-cmp" then
             local cmp = require("cmp")
 
             -- global sources
@@ -130,9 +130,7 @@ return {
     optional = true,
     opts = {
       sources = {
-        completion = {
-          enabled_providers = { "dadbod" },
-        },
+        default = { "dadbod" },
         providers = {
           dadbod = { name = "Dadbod", module = "vim_dadbod_completion.blink" },
         },