+++ lua/lazyvim/util/cmp.lua
@@ -13,6 +13,11 @@ M.actions = {
       return true
     end
   end,
+  snippet_stop = function()
+    if vim.snippet then
+      vim.snippet.stop()
+    end
+  end,
 }
 
 ---@param actions string[]
@@ -98,20 +103,6 @@ function M.add_missing_snippet_docs(window)
       end
     end
   end
-end
-
-function M.visible()
-  ---@module 'blink.cmp'
-  local blink = package.loaded["blink.cmp"]
-  if blink then
-    return blink.windows and blink.windows.autocomplete.win:is_open()
-  end
-  ---@module 'cmp'
-  local cmp = package.loaded["cmp"]
-  if cmp then
-    return cmp.core.view:visible()
-  end
-  return false
 end
 
 -- This is a better implementation of `cmp.confirm`: